WIP - classification algo 12/more refactor
